pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "flowcheck"
version = "0.3.0"
description = "A local MCP safety layer for developer workflow hygiene"
requires-python = ">=3.10"
authors = [
    { name = "FlowCheck Team", email = "maintainers@flowcheck.dev" }
]
license = "MIT"
keywords = ["mcp", "git", "developer-tools", "productivity", "security", "ai-safety", "agentic-workflow"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Version Control :: Git",
    "Topic :: Security",
    "Topic :: Software Development :: Quality Assurance",
]

dependencies = [
    "fastmcp>=2.0",
    "gitpython>=3.1",
]

[project.urls]
Homepage = "https://github.com/backslash-ux/flowcheck"
Documentation = "https://github.com/backslash-ux/flowcheck#readme"
Repository = "https://github.com/backslash-ux/flowcheck.git"
Issues = "https://github.com/backslash-ux/flowcheck/issues"
Changelog = "https://github.com/backslash-ux/flowcheck/blob/main/CHANGELOG.md"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio>=0.21",
]

[project.scripts]
flowcheck = "flowcheck.cli:main"
flowcheck-server = "flowcheck.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/flowcheck"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
